requirements.txt
boardgamegeek2==1.0.1
#Flask==1.1.2
Flask==2.0.1
Flask-Session
Flask-WTF
Flask-Bootstrap4
jinja2==3.0.1
py-bgg
redis
requests==2.3.0
requests_cache==0.4.13
